Move os::log to a separate header not requiring the os/(os).cpp include, which binds to a steinberg API. Rather inline the log implementation
