+++ cmake/wrap_standalone.cmake
@@ -184,14 +184,6 @@ function(target_add_standalone_wrapper)
                 ${CLAP_WRAPPER_CMAKE_CURRENT_SOURCE_DIR}/src/wrapasstandalone.cpp)
     endif()
 
-    if(WIN32)
-        target_sources(${SA_TARGET} PRIVATE ${CLAP_WRAPPER_CMAKE_CURRENT_SOURCE_DIR}/src/detail/os/windows.cpp)
-    elseif (APPLE)
-        target_sources(${SA_TARGET} PRIVATE ${CLAP_WRAPPER_CMAKE_CURRENT_SOURCE_DIR}/src/detail/os/macos.mm)
-    elseif(UNIX)
-        target_sources(${SA_TARGET} PRIVATE ${CLAP_WRAPPER_CMAKE_CURRENT_SOURCE_DIR}/src/detail/os/linux.cpp)
-    endif()
-
     if (DEFINED SA_HOSTED_CLAP_NAME)
         set(hasclapname TRUE)
     endif()
