+++ src/detail/os/windows.cpp
@@ -21,12 +21,6 @@ extern HINSTANCE ghInst;
 namespace os
 {
 
-void log(const char* text)
-{
-  OutputDebugStringA(text);
-  OutputDebugStringA("\n");
-}
-
 class WindowsHelper
 {
  public:
